pyproject.toml
[project]
name = "ohsome-filter-to-sql"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.4",
]

[tool.ruff]
extend-exclude = [
    "ohsome_filter_to_sql/OFLLexer.py",
    "ohsome_filter_to_sql/OFLListener.py",
    "ohsome_filter_to_sql/OFLParser.py"
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle Error
    "F",    # Pyflakes
    "I",    # isort
    "N",    # pep8-nameing
    "Q",    # flake8-quotes
    "W",    # pycodestyle Warning
    "C90",  # mccabe
    "TRY",  # tryceratops
    "RUF",  # ruff-specific rules
    "FURB", # refurb
    "PERF", # perflint
    "S",    # bandit
    "B",    # bugbear
    "SIM",  # simplify
]


[tool.ruff.lint.extend-per-file-ignores]
"tests/*" = ["S101"]  # ignore usage of asserts in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
